+++ application/lib/internship_organizer/period_manager/component/unsubscribe_group.class.php
@@ -11,9 +11,9 @@ class InternshipOrganizerPeriodManagerUnsubscribeGroupComponent extends Internsh
     function run()
     {
         $ids = Request :: get(InternshipOrganizerPeriodManager :: PARAM_PERIOD_REL_GROUP_ID);
-              
+        
         $failures = 0;
-               
+        
         if (! empty($ids))
         {
             if (! is_array($ids))
@@ -26,16 +26,23 @@ class InternshipOrganizerPeriodManagerUnsubscribeGroupComponent extends Internsh
                 
                 $period_rel_group_ids = explode('|', $id);
                 
-                $period_rel_group = InternshipOrganizerDataManager :: get_instance()->retrieve_period_rel_group($period_rel_group_ids[0], $period_rel_group_ids[1], $period_rel_group_ids[2]);
-                             
-                if (! $period_rel_group->delete())
+                $period_id = $period_rel_group_ids[0];
+                $location_id = InternshipOrganizerRights :: get_location_id_by_identifier_from_internship_organizers_subtree($period_id, InternshipOrganizerRights :: TYPE_PERIOD);
+                
+                if (InternshipOrganizerRights :: is_allowed_in_internship_organizers_subtree(InternshipOrganizerRights :: UNSUBSCRIBE_GROUP_RIGHT, $location_id, InternshipOrganizerRights :: TYPE_PERIOD))
                 {
-                    $failures ++;
-                }
-                else
-                {
-                                   	
-                	//                    Event :: trigger('delete', 'period', array('target_period_id' => $period->get_id(), 'action_user_id' => $user->get_id()));
+                    
+                    $period_rel_group = InternshipOrganizerDataManager :: get_instance()->retrieve_period_rel_group($period_rel_group_ids[0], $period_rel_group_ids[1], $period_rel_group_ids[2]);
+                    
+                    if (! $period_rel_group->delete())
+                    {
+                        $failures ++;
+                    }
+                    else
+                    {
+                        
+                    //                    Event :: trigger('delete', 'period', array('target_period_id' => $period->get_id(), 'action_user_id' => $user->get_id()));
+                    }
                 }
             }
             
@@ -61,7 +68,7 @@ class InternshipOrganizerPeriodManagerUnsubscribeGroupComponent extends Internsh
                     $message = 'SelectedInternshipOrganizerPeriodRelGroupsDeleted';
                 }
             }
-            $this->redirect(Translation :: get($message), ($failures ? true : false), array(InternshipOrganizerPeriodManager :: PARAM_ACTION => InternshipOrganizerPeriodManager :: ACTION_BROWSE_PERIODS, InternshipOrganizerPeriodManager :: PARAM_PERIOD_ID => $period_rel_group_ids[0], DynamicTabsRenderer::PARAM_SELECTED_TAB => InternshipOrganizerPeriodManagerBrowserComponent :: TAB_GROUPS));
+            $this->redirect(Translation :: get($message), ($failures ? true : false), array(InternshipOrganizerPeriodManager :: PARAM_ACTION => InternshipOrganizerPeriodManager :: ACTION_BROWSE_PERIODS, InternshipOrganizerPeriodManager :: PARAM_PERIOD_ID => $period_rel_group_ids[0], DynamicTabsRenderer :: PARAM_SELECTED_TAB => InternshipOrganizerPeriodManagerBrowserComponent :: TAB_GROUPS));
         }
         else
         {
